requirements.txt
alpaca-py==0.43.4
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
Flask==3.1.3
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.4.4
pandas==3.0.2
pydantic==2.13.4
pydantic_core==2.46.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
requests==2.33.1
six==1.17.0
sseclient-py==1.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
